Ingest HRV, rhythm, weight and blood sugar, and alert on all five
CardiTrack read 27 daily fields from the Google Health API and used eight
of them. This adds the five data types worth the most to a cardiac
product and were not read at all, alerts on each, and carries them
through every surface that already describes a member's day.

What is now ingested (field names and record types verified against the
v4 discovery document, revision 20260819):

- daily-heart-rate-variability — overnight RMSSD. Named as a secondary
  SSA series in docs/llm_design.md since the original design and never
  fetched; two designed alerts depended on it.
- electrocardiogram — the classification only. The waveform array is
  excluded by a fields partial-response selector and never crosses the
  wire: CardiTrack reports that a reading was taken and what the device
  made of it, which is what a family can act on.
- irregular-rhythm-notification — the device's own AFib screening.
- weight — the earliest signal of fluid retention a family can see.
- blood-glucose — listed rather than rolled up, because only the series
  carries the day's lowest reading and the low is what matters.

Five alert rules, in the settings catalogue under two new clusters
(Heart rhythm, Body & readings) and on by default:

- ecg_afib (red) and irregular_rhythm (orange) — new AlertType.Rhythm,
  kept out of HeartRate deliberately: that type is cooldown-scoped as a
  whole, so a standing "resting heart rate is running high" would have
  swallowed the most consequential alert in the product.
- hrv_drop (orange) — below baseline minus max(2σ, 15%) on both of the
  last two nights; one low night is a late meal, not a signal.
- rapid_weight_gain (orange) — 1.4 kg in three days or 2.3 kg in seven,
  the thresholds heart-failure self-monitoring guidance gives patients.
- blood_sugar_out_of_range — red below 70 mg/dL, orange above 250.

The alert engine now separates comparative rules, which need the
established 30-day baseline and stay silent without one, from measured
rules, which report a finding the device itself made. Gating the second
kind on a baseline would mean a member two weeks into wearing a watch
gets no word when it tells them their heart is in atrial fibrillation,
and there is no inference in that finding to be thin.

Carried through: daybook, weekbook and monthbook prompts, the shared
medical prompt block, digest computed observations and the refresh
waiver, dashboard metrics, baselines, chat charts and the device card's
dataset pills.

Also fixes a pre-existing bug in the same code path: both activity-log
upserts dropped TemperatureBaseline and TemperatureVariation when
updating an existing row, so the dashboard's temperature card could
compare a fresh reading against a stale baseline.

ECG and irregular-rhythm notifications need two new OAuth scopes
(googlehealth.ecg.readonly, googlehealth.irn.readonly). Connections
authorised before this ship without them and report null counts — never
zero — until the wearer reconnects, so neither rule can mistake "we
cannot see it" for "nothing happened".

+++ src/Infrastructure/CardiTrack.Infrastructure/Services/JournalPeriodSections.cs
@@ -55,7 +55,9 @@ internal static class JournalPeriodSections
         string label,
         Func<ActivityLog, T?> select,
         Func<decimal, string> format,
-        int? usual,
+        // decimal rather than int: HRV and weight keep two decimal places in the baseline, and
+        // widening here costs the integer baselines nothing (they convert implicitly).
+        decimal? usual,
         string? band,
         string periodNoun,
         Func<IReadOnlyList<(DateOnly Day, decimal Value)>, decimal, Func<decimal, string>, string?> standoutClause)
